move more common mapping to multiselect
